requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
cohere
numpy
